fix: PR review comments
