+++ source/javascripts/controllers/_MainController.js.erb
@@ -1,5 +1,6 @@
 import _ from "underscore";
 import datadogRumCustomTiming from "../utils/datadogCustomRumTiming.ts";
+import { safeDigest } from "../services/react-compat";
 
 (function() {
 	"use strict";
@@ -655,7 +656,7 @@ import datadogRumCustomTiming from "../utils/datadogCustomRumTiming.ts";
 				}
 
 				$rootScope.$emit("MainController::changesDiscarded");
-				$scope.$digest();
+				safeDigest($scope);
 			};
 
 			viewModel.selectMenuByID = function(menuID) {
