+++ source/templates/workflows.slim
@@ -54,14 +54,13 @@
 				.step-edit-container
 					/ div[ng-if="workflowsCtrl.selectedStep" style="background: white"]
 					/ 	r-step-config[
-					/ 		workflow-index="workflowIndex"
 					/ 		step="workflowsCtrl.selectedStep"
 					/ 		on-clone="workflowsCtrl.cloneSelectedStep"
 					/ 		on-remove="workflowsCtrl.deleteSelectedStep"
 					/ 		has-version-update="workflowsCtrl.checkStepVersionUpdate(workflowsCtrl.selectedStep)"
 					/ 		versions-with-remarks="workflowsCtrl.selectedStepVersionsWithRemarks"
 					/ 		output-variables="workflowsCtrl.outputVariables"
-					/ 		on-property-change="workflowsCtrl.handleStepPropertyChanged"
+					/ 		on-change="workflowsCtrl.handleStepPropertyChanged"
 					/ 	]
 					section.workflow
 						header.workflow-header
@@ -95,12 +94,11 @@
 											== svg("arrow-down")
 
 									r-step-version-details[
-										key="workflowsCtrl.selectedStep.$$hashKey"
-										step="workflowsCtrl.selectedStep"
-										latest-version="workflowsCtrl.selectedStepLatestVersion"
-										has-version-update="workflowsCtrl.checkStepVersionUpdate(workflowsCtrl.selectedStep)"
-										versions-with-remarks="workflowsCtrl.selectedStepVersionsWithRemarks"
-										on-property-change="workflowsCtrl.handleStepPropertyChanged"
+										step= "workflowsCtrl.selectedStep"
+										latest-version= "workflowsCtrl.selectedStepLatestVersion"
+										has-version-update= "workflowsCtrl.checkStepVersionUpdate(workflowsCtrl.selectedStep)"
+										versions-with-remarks= "workflowsCtrl.selectedStepVersionsWithRemarks"
+										on-change= "workflowsCtrl.handleStepPropertyChanged"
 									]
 
 									section.run-if
